requirements.txt
amqp==5.3.1
asgiref==3.8.1
attrs==25.3.0
billiard==4.2.1
cattrs==24.1.3
celery==5.5.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cron-descriptor==1.4.5
Django==5.2
django-celery-beat==2.8.0
django-cors-headers==4.7.0
django-db-connection-pool==1.2.5
django-debug-toolbar==5.1.0
django-filter==25.1
django-health-check==3.18.3
django-redis==5.4.0
django-rest-swagger==2.2.0
django-timezone-field==7.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.7
flatbuffers==25.2.10
greenlet==3.2.0
idna==3.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.6
kombu==5.5.3
MarkupSafe==3.0.2
numpy==2.2.4
openapi-codec==1.3.2
openmeteo_requests==1.4.0
openmeteo_sdk==1.20.0
packaging==24.2
pandas==2.2.3
platformdirs==4.3.7
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
requests-cache==1.2.1
retry-requests==2.0.0
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
setuptools==78.1.0
simplejson==3.20.1
six==1.17.0
SQLAlchemy==2.0.40
sqlparams==6.2.0
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
url-normalize==2.2.0
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13